backend/requirements.txt
feedparser==6.0.11
openai==1.54.3
firebase-admin==6.5.0
python-dotenv==1.0.1
requests==2.31.0
python-dateutil==2.9.0